feat: Extend pool stats and connection validation (Spec 023) (#32)
* feat: Extend pool stats and connection validation (Spec 023)

Implements 5 user stories for improved connection management:

- US1: TLS validation at ATTACH time with SELECT 1 query after authentication
- US2: Configurable query timeout via mssql_query_timeout setting (default 30s)
- US3: Catalog-free mode (catalog=false) for serverless/restricted databases
- US4: Extended pool stats with pinned_connections column
- US5: TLS enabled by default (use_encrypt=true) for improved security

Key changes:
- Add TLS-specific error messages for certificate/handshake failures
- Add query timeout with clear error messages and SET guidance
- Skip catalog initialization when catalog_enabled=false
- Track pinned connections in MSSQLTransaction for pool statistics
- Change default encryption from false to true (backward compatible)


* style: Fix clang-format violations

Apply clang-format to fix CI lint check failures in:
- src/mssql_storage.cpp
- src/query/mssql_query_executor.cpp
- src/query/mssql_result_stream.cpp
- src/connection/mssql_settings.cpp
- Related header files


* fix: Add missing includes and fix unique_ptr conversion in CTAS

- Add <cstdarg> and <cstdio> includes for va_start/va_end/fprintf
- Use std::move() for unique_ptr base class conversion in C++11 mode


* fix: Use explicit casts for format string portability in CTAS

Cast time values to (long long) with %lld format specifier for
consistent behavior across Linux (long int) and macOS (long long).


* docs: Update documentation for Spec 023 features

- Add query_timeout setting (mssql_query_timeout)
- Document TLS enabled by default (use_encrypt=true)
- Add catalog field for catalog-free mode
- Add pinned_connections to pool_stats output
- Document catalog-free mode for serverless databases
- Update PoolConfiguration and PoolStatistics structs


---------

+++ src/dml/ctas/mssql_ctas_executor.cpp
@@ -8,6 +8,8 @@
 #include "tds/tds_connection.hpp"
 #include "tds/tds_connection_pool.hpp"
 
+#include <cstdarg>
+#include <cstdio>
 #include <cstdlib>
 #include <cstring>
 
@@ -271,11 +273,11 @@ void CTASExecutionState::LogMetrics() const {
 	fprintf(stderr, "  Target: %s\n", target.GetQualifiedName().c_str());
 	fprintf(stderr, "  OR REPLACE: %s\n", target.or_replace ? "yes" : "no");
 	fprintf(stderr, "  DDL bytes: %llu\n", (unsigned long long)ddl_bytes);
-	fprintf(stderr, "  DDL time: %lld ms\n", ddl_time_ms);
+	fprintf(stderr, "  DDL time: %lld ms\n", (long long)ddl_time_ms);
 	fprintf(stderr, "  Rows produced: %llu\n", (unsigned long long)rows_produced);
 	fprintf(stderr, "  Rows inserted: %llu\n", (unsigned long long)rows_inserted);
-	fprintf(stderr, "  INSERT time: %lld ms\n", insert_time_ms);
-	fprintf(stderr, "  Total time: %lld ms\n", total_time_ms);
+	fprintf(stderr, "  INSERT time: %lld ms\n", (long long)insert_time_ms);
+	fprintf(stderr, "  Total time: %lld ms\n", (long long)total_time_ms);
 	fprintf(stderr, "  Phase: %s\n", GetPhaseName(phase).c_str());
 	if (!error_message.empty()) {
 		fprintf(stderr, "  Error: %s\n", error_message.c_str());
@@ -321,11 +323,11 @@ void CTASObservability::Log(int level) const {
 	fprintf(stderr, "  Target: %s\n", target_table.c_str());
 	fprintf(stderr, "  OR REPLACE: %s\n", or_replace ? "yes" : "no");
 	fprintf(stderr, "  DDL bytes: %llu\n", (unsigned long long)ddl_bytes);
-	fprintf(stderr, "  DDL time: %lld ms\n", ddl_time_ms);
+	fprintf(stderr, "  DDL time: %lld ms\n", (long long)ddl_time_ms);
 	fprintf(stderr, "  Rows produced: %llu\n", (unsigned long long)rows_produced);
 	fprintf(stderr, "  Rows inserted: %llu\n", (unsigned long long)rows_inserted);
 	fprintf(stderr, "  Batches: %llu\n", (unsigned long long)batches_executed);
-	fprintf(stderr, "  INSERT time: %lld ms\n", insert_time_ms);
+	fprintf(stderr, "  INSERT time: %lld ms\n", (long long)insert_time_ms);
 	fprintf(stderr, "  Success: %s\n", success ? "yes" : "no");
 	if (!success) {
 		fprintf(stderr, "  Failure phase: %s\n", failure_phase.c_str());
